Update HTML file with new structure and bio info
This commit updates the HTML of the index page by adding more detailed styles and a DOCTYPE at the top. It also replaces placeholder content with bio-boxes for Michael Hartl, Lee Donahoe, Nick Merwin and a placeholder for a future addition. Links have been slightly updated to include a trailing slash.